#1918: Fix a bunch of issues with pinned region detection.
Not every pinned region has a clean `P = null` assignment to mark its end.
If a second pinned region starts with the same variable `P`, consider that to mark the end of the previous pinned region for that variable.

Also, fix a bunch of special cases with empty pinned regions.